final changes

+++ app/Repositories/GameRepository.php
@@ -17,6 +17,11 @@ class GameRepository
 
     public function getGames(): Collection
     {
+        return Game::where('is_deleted',0)->orderBy('title')->get();
+    }
+
+    public function getGamesAdmin(): Collection
+    {
         return Game::orderBy('title')->get();
     }
 
@@ -44,17 +49,27 @@ class GameRepository
     }
 
     public function updateGame (int $game_id, string $title, int $publisher, array $category, int $release_year, string $description){
+        $message;
+        $alreadyExists = Game::firstWhere('title',$title);
         $timestamp = date("Y-m-d H:i:s");
-        DB::table('games')->where('id',$game_id)->update(['title'=>$title, 'description'=>$description, 'updated_at'=>$timestamp, 'release_year'=>$release_year,'publisher_id'=>$publisher]);
-        DB::table('category_game')->where('game_id',$game_id)->delete();
-        foreach($category as $value){
-            DB::insert('insert into category_game(game_id,category_id) values (?,?)',[$game_id, $value]);
+        if($alreadyExists == null || $alreadyExists->id == $game_id){
+            DB::table('games')->where('id',$game_id)->update(['title'=>$title, 'description'=>$description, 'updated_at'=>$timestamp, 'release_year'=>$release_year,'publisher_id'=>$publisher]);
+            DB::table('category_game')->where('game_id',$game_id)->delete();
+            foreach($category as $value){
+                DB::insert('insert into category_game(game_id,category_id) values (?,?)',[$game_id, $value]);
+            }
+            $message ="Game successfully added to database";
+            return $message;
+        }
+        else{
+            return $message = "This game title already exists in the database";
         }
     }
 
     public function addGame(string $title, int $publisher, array $category, int $release_year, string $description, $image = null){
         $message;
         $timestamp = date("Y-m-d H:i:s");
+
         $alreadyExists = Game::firstWhere('title',$title);
         if($alreadyExists == null){
             DB::insert('insert into games (title, image, description, created_at, updated_at, release_year, publisher_id, is_deleted) values (?,?,?,?,?,?,?,?)',
@@ -63,12 +78,12 @@ class GameRepository
             foreach($category as $value){
                 DB::insert('insert into category_game(game_id,category_id) values (?,?)',[$game->id, $value]);
             }
-                $message ="Game successfully added to database";
-            // return $message;
-        }
+            $message ="Game successfully added to database";
+            return $message;
+    }
         else{
             $message = "This game title already exists in the database";
-            // return $message;
+            return $message;
         }
     }
